Init commit.
General documentation:
All software on the server lives in /apps/cosc481
At server boot /etc/rc.local calls
/apps/cosc481/devops/server/runWebui.sh and runBackend.sh
These start the webui on port 8080 and the backend on port 9002. All
webpages from there are served by cherrypi.
When a request is received on port 80, nginx receives it, and does a
_local_ web request to localhost:8080, fiddles with the headers a bit,
and sends the response to the original requestor. Likewise port 9001 is
redirected to localhost port 9002
This is bad for performance and bad for security but makes development a
lot easier since we still get those sweet sweet cherrypi error messages
(stack traces and all!)
The infrastructure is in place to use uwsgi instead of local proxy
passes when we're stable enough to do that.
When a developer has made changes they would like to test they should
commit to github, then have the server clone their changes. For this a
script has been included: cosc481/devops/<your-platform>/updateServer
Web.py is usually very good about recompiling to include changes on the
fly, but some changes still require a reboot of the server -- some
changes will compile but cause cherrypi to be buggy.
Currently only Greg has reboot authority, but this might be changed if
it becomes a large enough problem.