Improve Look of Photos/Folders and upload/add cards
Added/rearranged divs, classes, ids related to the photo upload and
photo folder creation cards, as well as to the folders and photos
themselves.

Also updated rules for those classes/ids in the style.css rules

Also updated some photo and photoFolder functions in the ajax so that
it would get rid of most buggy behavior.

Also updated the templates at the bottom of job.html